add tests for definition analysis
