+++ fpy2/analysis/__init__.py
@@ -1,7 +1,7 @@
 """Program analyses for FPy programs"""
 
 from .context_infer import ContextInfer, ContextAnalysis, ContextInferError, TypeContext
-from .defs import DefinitionAnalysis
+from .defs import DefAnalysis
 from .define_use import (
     DefineUse, DefineUseAnalysis, Definition, DefinitionCtx,
     AssignDef, PhiDef,
