+++ tests/infra/unit/dead_code.py
@@ -4,7 +4,7 @@ Tests for dead code elimination.
 
 import fpy2 as fp
 
-from .defs import tests, examples
+from .unit_tests import tests, examples
 
 _modules = [
     fp.libraries.core,
