+++ tests/infra/unit/eval.py
@@ -1,4 +1,4 @@
-from .defs import tests, examples
+from .unit_tests import tests, examples
 
 def test_eval():
     for core in tests + examples:
